refactored start wizard
